pyproject.toml
[project]
name = "teg-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.0",
]

[tool.uv.workspace]
members = [
    "backend",
    "frontend",
    "ai",
]
